alice v0.1.1 Alice.Handlers.Help

Functions

commands()

All of the commands handled by this module

delete_state(conn, key)

Deletes the entries in the state for a specific key.

get_state(conn, key, default \\ nil)

Get the state from an Alice.Conn struct, namespaced to this module

handle(conn, name)

Callback. This function gets called when a route with name matches

Must return an Alice.Conn.

match_commands(conn)

Match all commands in this module

match_routes(conn)

Match all routes in this module

max_length(routes)

put_state(conn, key, value)

Update the state of an Alice.Conn struct, namespaced to this module

routes()

All of the routes handled by this module