alice v0.1.1 Alice.Router.Helpers

Functions

chance_reply(chance, positive, negative \\ :noreply, conn)

Reply with random chance.

Examples

> chance_reply(0.5, "this will be sent half the time, otherwise nothing will be sent")
> chance_reply(0.25, "this will be sent 25% of the time", "sent 75% of the time")

random_reply(list, conn)

Replies with a random element of the list provided.

reply(response, map)

Reply to a message in a handler.

Sends response back to the channel that triggered the handler.

Adds random tag to end of image urls to break Slack’s img cache.

Macros

command(pattern, name)

Adds a command to the handler

route(pattern, name)

Adds a route to the handler